+++ laktory/cli/__init__.py
@@ -1,13 +1,14 @@
 from laktory.cli.app import app
-import laktory.cli._init
 import laktory.cli._deploy
+import laktory.cli._destroy
+import laktory.cli._init
 import laktory.cli._preview
 import laktory.cli._quickstart
 import laktory.cli._run
-import laktory.cli._write
 
-from laktory.cli._init import init
 from laktory.cli._deploy import deploy
+from laktory.cli._destroy import destroy
+from laktory.cli._init import init
 from laktory.cli._preview import preview
 from laktory.cli._quickstart import quickstart
 from laktory.cli._run import run