+++ laktory/resources/quickstart-stacks/workflows/sql/slv_stock_prices.sql
@@ -1,0 +1,7 @@
+SELECT
+    cast(data.created_at AS TIMESTAMP) AS created_at,
+    data.symbol AS symbol,
+    data.open AS open,
+    data.close AS close
+FROM
+    {df}